requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0
django-rest-swagger==2.2.0
django-tinymce==2.8.0
djangorestframework==3.10.3
gunicorn==20.0.4
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==6.2.1
psycopg2==2.8.4
python-decouple==3.3
pytz==2019.3
requests==2.22.0
simplejson==3.17.0
sqlparse==0.3.0
supervisor==4.1.0
uritemplate==3.0.0
urllib3==1.25.7
